requirements.txt
# --- Core Framework ---
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# --- Integrations ---
langchain-ollama>=0.2.0
langchain-chroma>=0.1.4

# --- Document Loading ---
pypdf

# --- User Interface ---
gradio>=4.0.0

# --- Utilities & Tracing ---
python-dotenv
langsmith